pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pyscroll-ce"
version = "2.0"
description = "Fast scrolling maps library for pygame"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Giorgio", email = "menzigi@duck.com" }
]
requires-python = ">=3.10"

classifiers = [
    "Intended Audience :: Developers",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: pygame",
]

dependencies = [
    "pygame-ce",
    "pytmx",
]

[project.urls]
source = "https://github.com/JaskRendix/pyscroll-ce"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["pyscroll"]

[tool.ruff]
line-length = 88
target-version = "py310"
src = ["pyscroll"]

[tool.ruff.lint]
extend-select = ["I"]  # enable import sorting

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
